# VertexPath.CalculateClosestPointOnPathData optimization