--- a/Path Creator Project/Assets/PathCreator/Core/Runtime/Objects/VertexPath.cs
+++ b/Path Creator Project/Assets/PathCreator/Core/Runtime/Objects/VertexPath.cs
@@ -32,7 +32,7 @@ namespace PathCreation {
         /// Equal to (0,0,-1) for 2D paths, and (0,1,0) for XZ paths
         public readonly Vector3 up;
 
-        // Default values and constants:    
+        // Default values and constants:
         const int accuracy = 10; // A scalar for how many times bezier path is divided when determining vertex positions
         const float minVertexSpacing = .01f;
 
@@ -228,19 +228,29 @@ namespace PathCreation {
 
         /// Finds the closest point on the path from any point in the world
         public Vector3 GetClosestPointOnPath (Vector3 worldPoint) {
-            TimeOnPathData data = CalculateClosestPointOnPathData (worldPoint);
-            return Vector3.Lerp (GetPoint (data.previousIndex), GetPoint (data.nextIndex), data.percentBetweenIndices);
+            // Transform the provided worldPoint into VertexPath local-space.
+            // This allows to do math on the localPoint's, thus avoiding the need to
+            // transform each local vertexpath point into world space via GetPoint.
+            Vector3 localPoint = MathUtility.InverseTransformPoint(worldPoint, transform, space);
+
+            TimeOnPathData data = CalculateClosestPointOnPathData (localPoint);
+            Vector3 localResult = Vector3.Lerp (localPoints[data.previousIndex], localPoints[data.nextIndex], data.percentBetweenIndices);
+
+            // Transform local result into world space
+            return MathUtility.TransformPoint(localResult, transform, space);
         }
 
         /// Finds the 'time' (0=start of path, 1=end of path) along the path that is closest to the given point
         public float GetClosestTimeOnPath (Vector3 worldPoint) {
-            TimeOnPathData data = CalculateClosestPointOnPathData (worldPoint);
+            Vector3 localPoint = MathUtility.InverseTransformPoint(worldPoint, transform, space);
+            TimeOnPathData data = CalculateClosestPointOnPathData (localPoint);
             return Mathf.Lerp (times[data.previousIndex], times[data.nextIndex], data.percentBetweenIndices);
         }
 
         /// Finds the distance along the path that is closest to the given point
         public float GetClosestDistanceAlongPath (Vector3 worldPoint) {
-            TimeOnPathData data = CalculateClosestPointOnPathData (worldPoint);
+            Vector3 localPoint = MathUtility.InverseTransformPoint(worldPoint, transform, space);
+            TimeOnPathData data = CalculateClosestPointOnPathData(localPoint);
             return Mathf.Lerp (cumulativeLengthAtEachVertex[data.previousIndex], cumulativeLengthAtEachVertex[data.nextIndex], data.percentBetweenIndices);
         }
 
@@ -248,7 +258,7 @@ namespace PathCreation {
 
         #region Internal methods
 
-        /// For a given value 't' between 0 and 1, calculate the indices of the two vertices before and after t. 
+        /// For a given value 't' between 0 and 1, calculate the indices of the two vertices before and after t.
         /// Also calculate how far t is between those two vertices as a percentage between 0 and 1.
         TimeOnPathData CalculatePercentOnPathData (float t, EndOfPathInstruction endOfPathInstruction) {
             // Constrain t based on the end of path instruction
@@ -295,7 +305,7 @@ namespace PathCreation {
         }
 
         /// Calculate time data for closest point on the path from given world point
-        TimeOnPathData CalculateClosestPointOnPathData (Vector3 worldPoint) {
+        TimeOnPathData CalculateClosestPointOnPathData (Vector3 localPoint) {
             float minSqrDst = float.MaxValue;
             Vector3 closestPoint = Vector3.zero;
             int closestSegmentIndexA = 0;
@@ -311,8 +321,8 @@ namespace PathCreation {
                     }
                 }
 
-                Vector3 closestPointOnSegment = MathUtility.ClosestPointOnLineSegment (worldPoint, GetPoint (i), GetPoint (nextI));
-                float sqrDst = (worldPoint - closestPointOnSegment).sqrMagnitude;
+                Vector3 closestPointOnSegment = MathUtility.ClosestPointOnLineSegment (localPoint, localPoints[i], localPoints[nextI]);
+                float sqrDst = (localPoint - closestPointOnSegment).sqrMagnitude;
                 if (sqrDst < minSqrDst) {
                     minSqrDst = sqrDst;
                     closestPoint = closestPointOnSegment;
@@ -321,8 +331,8 @@ namespace PathCreation {
                 }
 
             }
-            float closestSegmentLength = (GetPoint (closestSegmentIndexA) - GetPoint (closestSegmentIndexB)).magnitude;
-            float t = (closestPoint - GetPoint (closestSegmentIndexA)).magnitude / closestSegmentLength;
+            float closestSegmentLength = (localPoints[closestSegmentIndexA] - localPoints[closestSegmentIndexB]).magnitude;
+            float t = (closestPoint - localPoints[closestSegmentIndexA]).magnitude / closestSegmentLength;
             return new TimeOnPathData (closestSegmentIndexA, closestSegmentIndexB, t);
         }