--- a/Path Creator Project/Assets/PathCreator/Examples/Scenes/GetClosestPointOnPath Test.unity.meta
+++ b/Path Creator Project/Assets/PathCreator/Examples/Scenes/GetClosestPointOnPath Test.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fbb1f45e5065be44badcc1594a52435a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: