--- a/Path Creator Project/Assets/PathCreator/Examples/Scripts/GetClosestPointOnPathTest.cs.meta
+++ b/Path Creator Project/Assets/PathCreator/Examples/Scripts/GetClosestPointOnPathTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2353903649ab3c34eb2a2bffc094de57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: