This PR contains an optimized version of VertexPath.CalculateClosestPointOnPathData, which positively affects the performance of:

- VertexPath.GetClosestPointOnPath
- VertexPath.GetClosestTimeOnPath
- VertexPath.GetClosestDistanceAlongPath

The idea of the optimization is to perform all the math and specifically "ClosestPointOnLineSegment" in local-space, because then there is no need to transform every point in the path to world-space.

This works by transforming the incoming "worldPoint" from world-space to "VertexPath local space" only once. The result CalculateClosestPointOnPathData result is then transformed from local-  to world-space again.

On my machine, profiling before and after the optimization shows:

- Old implementation: 5.58ms - 14.62ms
- New implementation: 0.25ms - 0.40ms

I've used the scene "GetClosestPointOnPath Test" to test and profile the new implementation against the old one.