@SebLague 2020-02-17
Great optimization, thanks!